+++ example/lib/ui/screens/mdpSQLite/MDPSQLiteScreen.dart
@@ -8,7 +8,9 @@ import 'package:example/ui/screens/mdpSQLite/MDPSQLiteRecordScreen.dart';
 import 'package:tch_appliable_core/tch_appliable_core.dart';
 
 class MDPSQLiteScreen extends AbstractAppScreen {
-  static const String ROUTE = "/mdpsqlite";
+  static const String ROUTE = '/mdpsqlite';
+
+  MDPSQLiteScreen({super.key});
 
   /// Create state for widget
   @override
@@ -23,7 +25,7 @@ class _MDPSQLiteScreenState extends AbstractAppScreenState<MDPSQLiteScreen> {
   )..appBarOptions = <AppBarOption>[
       AppBarOption(
         onTap: (BuildContext context) {
-          pushNamed(context, MDPSQLiteRecordScreen.ROUTE);
+          pushNamedV2(context, MDPSQLiteRecordScreen.ROUTE);
         },
         icon: Icon(
           Icons.add,
@@ -67,9 +69,6 @@ class _MDPSQLiteScreenState extends AbstractAppScreenState<MDPSQLiteScreen> {
 }
 
 class _BodyWidget extends AbstractStatefulWidget {
-  /// BodyWidget initialization
-  _BodyWidget({Key? key}) : super(key: key);
-
   /// Create state for widget
   @override
   State<StatefulWidget> createState() => _BodyWidgetState();
@@ -90,15 +89,17 @@ class _BodyWidgetState extends AbstractStatefulWidgetState<_BodyWidget> {
         return Material(
           color: Colors.transparent,
           child: InkWell(
-            child: Container(
+            child: SizedBox(
               height: 48,
-              padding: const EdgeInsets.symmetric(horizontal: 16),
-              child: Row(
-                mainAxisSize: MainAxisSize.max,
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: [
-                  Flexible(child: Text(item.name)),
-                ],
+              child: Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 16),
+                child: Row(
+                  mainAxisSize: MainAxisSize.max,
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: [
+                    Flexible(child: Text(item.name)),
+                  ],
+                ),
               ),
             ),
             onLongPress: () {
@@ -113,10 +114,12 @@ class _BodyWidgetState extends AbstractStatefulWidgetState<_BodyWidget> {
           text: Text(tt('list.item.loading')),
         );
       },
-      emptyState: Container(
+      emptyState: SizedBox(
         width: 576,
-        padding: const EdgeInsets.all(16),
-        child: Text(tt('list.empty')),
+        child: Padding(
+          padding: const EdgeInsets.all(16),
+          child: Text(tt('list.empty')),
+        ),
       ),
     );
   }