+++ lib/src/ui/screens/abstract_responsive_screen.dart
@@ -3,7 +3,9 @@ import 'package:tch_appliable_core/src/core/core_app.dart';
 import 'package:tch_appliable_core/src/ui/screens/abstract_screen.dart';
 import 'package:tch_appliable_core/src/ui/widgets/abstract_responsive_widget.dart';
 
-abstract class AbstractResponsiveScreen extends AbstractScreen {}
+abstract class AbstractResponsiveScreen extends AbstractScreen {
+  AbstractResponsiveScreen({super.key});
+}
 
 abstract class AbstractResponsiveScreenState<T extends AbstractResponsiveScreen> extends AbstractScreenState<T> {
   /// Create view layout from widgets for phone screens (< 576px)
@@ -37,22 +39,22 @@ abstract class AbstractResponsiveScreenState<T extends AbstractResponsiveScreen>
 
     if (snapshot != null) {
       switch (snapshot.responsiveScreen) {
-        case ResponsiveScreen.ExtraLargeDesktop:
+        case ResponsiveScreen.extraLargeDesktop:
           return extraLargeDesktopScreen(context);
-        case ResponsiveScreen.LargeDesktop:
+        case ResponsiveScreen.largeDesktop:
           return largeDesktopScreen(context);
-        case ResponsiveScreen.SmallDesktop:
+        case ResponsiveScreen.smallDesktop:
           return smallDesktopScreen(context);
-        case ResponsiveScreen.Tablet:
+        case ResponsiveScreen.tablet:
           return tabletScreen(context);
-        case ResponsiveScreen.LargePhone:
+        case ResponsiveScreen.largePhone:
           return largePhoneScreen(context);
-        case ResponsiveScreen.SmallPhone:
-        case ResponsiveScreen.UnDetermined:
+        case ResponsiveScreen.smallPhone:
+        case ResponsiveScreen.unDetermined:
           return smallPhoneScreen(context);
       }
     }
 
-    throw Exception('Snapshot not available');
+    return smallPhoneScreen(context);
   }
 }